refactor(aircall): implement common rate limit middleware
