+++ apps/aircall/src/connectors/aircall/auth.test.ts
@@ -2,7 +2,7 @@ import { http } from 'msw';
 import { describe, expect, test, beforeEach } from 'vitest';
 import { server } from '@elba-security/test-utils';
 import { env } from '@/common/env';
-import { AircallError } from '../common/error';
+import { AircallError } from './common/error';
 import { getToken } from './auth';
 
 const validCode = '1234';
