+++ apps/aircall/src/connectors/aircall/common/error.ts
