+++ apps/aircall/src/inngest/client.ts
@@ -1,6 +1,7 @@
 import { EventSchemas, Inngest } from 'inngest';
 import { logger } from '@elba-security/logger';
-import { rateLimitMiddleware } from './middlewares/rate-limit-middleware';
+import { rateLimitMiddleware } from '@elba-security/inngest';
+import { getRetryAfter } from '@/connectors/aircall/common/retry-after';
 import { unauthorizedMiddleware } from './middlewares/unauthorized-middleware';
 
 export const inngest = new Inngest({
@@ -37,6 +38,6 @@ export const inngest = new Inngest({
       };
     };
   }>(),
-  middleware: [rateLimitMiddleware, unauthorizedMiddleware],
+  middleware: [rateLimitMiddleware({ getRetryAfter }), unauthorizedMiddleware],
   logger,
 });
