--- apps/aircall/src/inngest/middlewares/rate-limit-middleware.test.ts
@@ -1,81 +1,0 @@
-import { afterAll, beforeAll, describe, expect, test, vi } from 'vitest';
-import { RetryAfterError } from 'inngest';
-import { AircallError } from '@/connectors/common/error';
-import { rateLimitMiddleware } from './rate-limit-middleware';
-
-const now = new Date('2024-06-17T10:00:00.000Z');
-
-describe('rate-limit middleware', () => {
-  beforeAll(() => {
-    vi.setSystemTime(now);
-  });
-
-  afterAll(() => {
-    vi.useRealTimers();
-  });
-
-  test('should not transform the output when their is no error', () => {
-    expect(
-      rateLimitMiddleware
-        .init()
-        // @ts-expect-error -- this is a mock
-        .onFunctionRun({ fn: { name: 'foo' } })
-        .transformOutput({
-          result: {},
-        })
-    ).toBeUndefined();
-  });
-
-  test('should not transform the output when the error is not about Aircall rate limit', () => {
-    expect(
-      rateLimitMiddleware
-        .init()
-        // @ts-expect-error -- this is a mock
-        .onFunctionRun({ fn: { name: 'foo' } })
-        .transformOutput({
-          result: {
-            error: new Error('foo bar'),
-          },
-        })
-    ).toBeUndefined();
-  });
-
-  test('should transform the output error to RetryAfterError when the error is about aircall rate limit', () => {
-    vi.setSystemTime(now);
-    const rateLimitError = new AircallError('foo bar', {
-      // @ts-expect-error this is a mock
-      response: {
-        status: 429,
-        headers: new Headers({ 'x-aircallapi-reset': String(now.getTime() + 180000) }), // 3 minutes
-      },
-    });
-
-    const context = {
-      foo: 'bar',
-      baz: {
-        biz: true,
-      },
-      result: {
-        data: 'bizz',
-        error: rateLimitError,
-      },
-    };
-
-    const result = rateLimitMiddleware
-      .init()
-      // @ts-expect-error -- this is a mock
-      .onFunctionRun({ fn: { name: 'foo' } })
-      .transformOutput(context);
-    expect(result?.result.error).toBeInstanceOf(RetryAfterError);
-    expect(result?.result.error.retryAfter).toStrictEqual('180');
-    expect(result).toMatchObject({
-      foo: 'bar',
-      baz: {
-        biz: true,
-      },
-      result: {
-        data: 'bizz',
-      },
-    });
-  });
-});
